# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AttachmentDto": {
        "properties": {
          "byteSize": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": ["id", "filename", "byteSize", "createdAt"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "crm_session": {
        "in": "cookie",
        "name": "crm_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "多租户 CRM 平台 REST API",
    "title": "Multi-tenant CRM API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces/{tenantCode}/objects/{objectCode}/records/{recordId}/attachments": {
      "get": {
        "operationId": "AttachmentsController_list",
        "parameters": [
          {
            "in": "path",
            "name": "objectCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recordId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttachmentDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["attachments"]
      },
      "post": {
        "operationId": "AttachmentsController_upload",
        "parameters": [
          {
            "in": "path",
            "name": "objectCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recordId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": ["file"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["attachments"]
      }
    },
    "/api/v1/workspaces/{tenantCode}/objects/{objectCode}/records/{recordId}/attachments/{id}": {
      "delete": {
        "operationId": "AttachmentsController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "objectCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recordId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["attachments"]
      },
      "get": {
        "operationId": "AttachmentsController_download",
        "parameters": [
          {
            "in": "path",
            "name": "objectCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recordId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["attachments"]
      }
    },
...
  },
  "servers": [],
...
}